Allow Customized LDB to open CFs with different comparators with "--try_load_options"
Summary: Right now even customized LDB cannot open a DB with CFs of different comparator (unless number of column families are pre-defined to the tool), even if all the comparators used are of known types. "--try_load_options" doesn't work as when loading options, comparators are not set. This commit allow users to specify a comparator name to comparator map, so that options can be constructed with the correct comparators.

Test Plan: Add a unit test.